Add files via upload
We have completed project 1 of sprint 1 at TripleTen (formerly known as Practicum), in which three hypotheses related to user activity and musical preferences on the platform were evaluated.

During the analysis, it was found that user activity at TripleTen varies depending on the day of the week, although differences between participating cities were also observed. The first hypothesis was fully accepted, indicating a significant relationship between user activity and the day of the week in different cities.

However, when examining musical preferences in relation to the days of the week, it was determined that there were no significant variations in genre choices on TripleTen. Although small differences were observed in Monday's records, no significant changes were found throughout the week. Therefore, the second hypothesis was not accepted.

Regarding the third hypothesis, it was found that the musical preferences of TripleTen users were similar across all participating cities. While the possibility of differences was considered, there was insufficient evidence in the data to support this hypothesis. Therefore, the third hypothesis was also rejected.

It is important to note that this project is just a sample of what can be achieved at TripleTen. In future projects, more precise and quantitative analyses will be conducted, focusing on the study of statistical hypotheses. Additionally, efforts will be made to expand the scope of projects to obtain more representative data and generalize the results on a platform-wide scale.

The TripleTen team is excited about the results obtained in this project and looks forward to continuing to explore new hypotheses and challenges in future data analysis sprints.