How Family Factors Impact Loan Repayments
Understanding Loan Default Factors in Banking

This study aims to discern whether specific variables, such as marital status and the number of children, play a role in a client's ability to repay a loan. By meticulously examining data provided by a banking institution regarding client solvency, we've pinpointed trends and patterns related to loan default. The analysis also addresses data quality issues, like missing values and duplicates, ensuring findings are accurate and reliable. The insights derived from this research are anticipated to be instrumental in the future assessment of a potential client's creditworthiness, assisting the bank in making well-informed lending decisions. It's crucial to emphasize that for a comprehensive understanding of default factors, a more extensive and context-specific data collection aligned with the particular reality of the financial institution is recommended.